Refactor styles for input and note components; add new styles for recipe header and directions
