+++ src/components/directions/DirectionsList.module.css
@@ -1,0 +1,17 @@
+.direction {
+  margin-top: 0;
+  display: flex;
+  flex-direction: row;
+  gap: 10px;
+}
+
+.direction-input {
+  width: 100%;
+  padding: 0;
+}
+
+.directions {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
