+++ src/components/image-uploader/ImageUploader.module.css
@@ -1,0 +1,61 @@
+.image-container {
+  height: 250px;
+  width: 250px;
+  border: 1px solid white;
+  position: relative;
+}
+
+.image-container:hover .image-overlay {
+  opacity: 1;
+}
+
+.image {
+  position: absolute;
+  top: 0;
+  left: 0;
+  height: 250px;
+  width: 250px;
+  object-fit: cover;
+}
+
+.image-overlay {
+  height: 250px;
+  width: 250px;
+  opacity: 0;
+}
+
+.upload-icon {
+  height: 25px;
+  width: 25px;
+  opacity: 0.6;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+}
+
+[data-theme="light"] .upload-icon {
+  filter: var(--icon-filter);
+}
+
+.upload-icon:hover {
+  cursor: pointer;
+}
+
+.close-icon {
+  position: absolute;
+  top: 10px;
+  right: 10px;
+  height: 25px;
+  width: 25px;
+  opacity: 0.6;
+  filter: drop-shadow(0px 0px 2px black);
+}
+
+[data-theme="light"] .close-icon {
+  filter: var(--icon-filter);
+}
+
+.close-icon:hover {
+  cursor: pointer;
+}
