+++ src/components/ingredients/IngredientsList.module.css
@@ -1,0 +1,94 @@
+.ingredient {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  margin-top: -7px;
+}
+
+.ingredient-measure {
+  font-weight: bold;
+  margin-left: 10px;
+}
+
+.ingredient-measure input {
+  height: 30px;
+}
+
+.ingredient-measure input::placeholder {
+  font-style: italic;
+}
+
+.ingredient-unit {
+  margin-left: 3px;
+}
+
+.ingredient-name {
+  margin-left: 3px;
+  width: 100%;
+  margin-top: 4px;
+}
+
+.ingredient-name::placeholder {
+  font-style: italic;
+}
+
+.ingredients-header {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+}
+
+.unit-toggle {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+}
+
+.switch {
+  position: relative;
+  display: inline-block;
+  width: 40px;
+  height: 20px;
+}
+
+.switch input {
+  opacity: 0;
+  width: 0;
+  height: 0;
+}
+
+.slider {
+  position: absolute;
+  cursor: pointer;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: #ccc;
+  transition: .4s;
+  border-radius: 20px;
+}
+
+.slider:before {
+  position: absolute;
+  content: "";
+  height: 16px;
+  width: 16px;
+  left: 2px;
+  bottom: 2px;
+  background-color: white;
+  transition: .4s;
+  border-radius: 50%;
+}
+
+input:checked + .slider {
+  background-color: #2196F3;
+}
+
+input:checked + .slider:before {
+  transform: translateX(20px);
+}
+
+.unit {
+  font-style: italic;
+}
