+++ src/components/recipe-header/RecipeHeader.module.scss
@@ -1,0 +1,12 @@
+
+.title-input {
+  font-size: 26px !important;
+  font-weight: bold !important;
+  margin-bottom: 10px;
+}
+
+.description-input {
+  margin-top: 20px;
+  margin-bottom: 10px;
+  width: 80%;
+}
