+++ src/components/view-editor/ViewEditor.module.css
@@ -58,209 +58,3 @@
   flex-direction: column;
 }
 
-.image-container {
-  height: 250px;
-  width: 250px;
-  border: 1px solid white;
-  position: relative;
-}
-
-.image-container:hover .image-overlay {
-  opacity: 1;
-}
-
-.image {
-  position: absolute;
-  top: 0;
-  left: 0;
-  height: 250px;
-  width: 250px;
-  object-fit: cover;
-}
-
-.image-overlay {
-  height: 250px;
-  width: 250px;
-  opacity: 0;
-}
-
-.close-icon {
-  position: absolute;
-  top: 10px;
-  right: 10px;
-  height: 25px;
-  width: 25px;
-  opacity: 0.6;
-  filter: drop-shadow(0px 0px 2px black);
-}
-
-[data-theme="light"] .close-icon {
-  filter: var(--icon-filter);
-}
-
-.close-icon:hover {
-  cursor: pointer;
-}
-
-.ingredient {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  margin-top: -7px;
-}
-
-.ingredient-measure {
-  font-weight: bold;
-  margin-left: 10px;
-}
-
-.ingredient-measure input {
-  height: 30px;
-}
-
-.ingredient-measure input::placeholder {
-  font-style: italic;
-}
-
-.ingredient-unit {
-  margin-left: 3px;
-}
-
-.ingredient-name {
-  margin-left: 3px;
-  width: 100%;
-  margin-top: 4px;
-}
-
-.ingredient-name::placeholder {
-  font-style: italic;
-}
-
-.direction {
-  margin-top: 0;
-  display: flex;
-  flex-direction: row;
-  gap: 10px;
-}
-
-.note {
-  margin-top: 0;
-  display: flex;
-  flex-direction: row;
-  gap: 10px;
-}
-
-.note-input {
-  width: 100%;
-}
-
-input {
-  text-align: left;
-  color: var(--text-primary);
-}
-
-input:focus {
-  outline: none;
-}
-
-.title-input {
-  font-size: 26px !important;
-  font-weight: bold !important;
-  margin-bottom: 10px;
-}
-
-.description-input {
-  margin-top: 20px;
-  margin-bottom: 10px;
-  width: 80%;
-}
-
-.upload-icon {
-  height: 25px;
-  width: 25px;
-  opacity: 0.6;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
-
-[data-theme="light"] .upload-icon {
-  filter: var(--icon-filter);
-}
-
-.upload-icon:hover {
-  cursor: pointer;
-}
-
-.direction-input {
-  width: 100%;
-  padding: 0;
-}
-
-.directions {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.ingredients-header {
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-}
-
-.unit-toggle {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.switch {
-  position: relative;
-  display: inline-block;
-  width: 40px;
-  height: 20px;
-}
-
-.switch input {
-  opacity: 0;
-  width: 0;
-  height: 0;
-}
-
-.slider {
-  position: absolute;
-  cursor: pointer;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: #ccc;
-  transition: .4s;
-  border-radius: 20px;
-}
-
-.slider:before {
-  position: absolute;
-  content: "";
-  height: 16px;
-  width: 16px;
-  left: 2px;
-  bottom: 2px;
-  background-color: white;
-  transition: .4s;
-  border-radius: 50%;
-}
-
-input:checked + .slider {
-  background-color: #2196F3;
-}
-
-input:checked + .slider:before {
-  transform: translateX(20px);
-}
-
-.unit {
-  font-style: italic;
-}
